Return Category in CategoryRepository

+++ RookieEShop/RookieEShop.BackEnd/Repositories/CategoryRepository.cs
@@ -18,12 +18,12 @@ namespace RookieEShop.BackEnd.Repositories
 			_dbContext = dbContext;
 		}
 
-		public async Task<IEnumerable<CategoryVm>> ReadAllCategory()
+		public async Task<IEnumerable<Category>> ReadAllCategory()
 		{
 			return await _dbContext.Categories.ToListAsync();
 		}
 
-		public async Task<CategoryVm> ReadDetailCategory(int id)
+		public async Task<Category> ReadDetailCategory(int id)
 		{
 			return await _dbContext.Categories.FindAsync(id);
 		}
@@ -47,6 +47,11 @@ namespace RookieEShop.BackEnd.Repositories
 		{
 			var oldCategory = await _dbContext.Categories.FindAsync(id);
 
+			if(oldCategory == null)
+			{
+				return false;
+			}
+
 			oldCategory.Name = category.Name;
 
 			if (await _dbContext.SaveChangesAsync() > 0)
@@ -63,6 +68,11 @@ namespace RookieEShop.BackEnd.Repositories
 		public async Task<bool> DeleteCategory(int id)
 		{
 			var oldCategory = await _dbContext.Categories.FindAsync(id);
+
+			if(oldCategory == null)
+			{
+				return false;
+			}
 
 			_dbContext.Categories.Remove(oldCategory);
 			if(await _dbContext.SaveChangesAsync() > 0)
